pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | migrations
    | config/settings
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.poetry]
name = "lubimovka_backend"
version = "0.1.0"
description = "Lubimovka + Ya.Practicum backend"
authors = ["Ya.Practicum Students"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-environ = "^0.7.0"
django-cors-headers = "^3.8.0"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
gunicorn = "^20.1.0"
drf-spectacular = "^0.18.2"
django-filter = "^2.4.0"
Pillow = "^8.3.2"
django-gfklookupwidget = "^1.0.9"
django-admin-sortable2 = "^1.0.2"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
pre-commit = "^2.15.0"
isort = "^5.9.3"
flake8 = "^3.9.2"
black = "^21.9b0"
ipython = "^7.27.0"
pytest-django = "^4.4.0"
coverage = "^5.5"
Werkzeug = "^2.0.1"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
